Applied modifications for map bottom sheet's view (abbreviations)

+++ app/src/main/java/com/emplk/realestatemanager/domain/property/type_price_surface/PropertyTypePriceAndSurfaceEntity.kt
@@ -6,7 +6,7 @@ data class PropertyTypePriceAndSurfaceEntity(
     val id: Long,
     val type: String,
     val price: BigDecimal,
-    val surface: Int,
+    val surface: Double,
     val featuredPictureUri: String,
     val pictureUri: String,
     //   val amenities: List<AmenityEntity>,
